requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==5.1.3
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
Faker==33.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
